Change JS to TS in backend.
