+++ backend/.dockerignore
@@ -2,3 +2,4 @@
 Dockerfile
 docker-compose
 node_modules
+dist
