+++ backend/Dockerfile
@@ -9,6 +9,8 @@ RUN npm install
 
 COPY . .
 
+RUN npm run build
+
 EXPOSE 5000
 
 CMD ["npm", "start"]
