+++ backend/package.json
@@ -1,7 +1,15 @@
 {
+  "scripts": {
+    "start": "node dist/server.js",
+    "build": "tsc"
+  },
   "dependencies": {
     "express": "^4.18.2",
     "mongodb": "^5.6.0",
-    "mongoose": "^7.2.3"
+    "mongoose": "^7.2.3",
+    "typescript": "^5.1.3"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.17"
   }
 }
